PICKit3 programming script (set up in "Tools"):
See: https://forum.mikroe.com/viewtopic.php?f=88&t=55432
Note that this script is different from that presented in that it
includes /MC to programme the configuration memory and /L to release
the device from reset, otherwise it won't run after programming.

In the MikroC Tools/Options menu make a new tools entry to run the batch
file using this parameter line (note the quotes):

%CHIP_NAME "%HEX_FILE_NAME"

The PK3CMD.exe cmd-line tool is installed along with MPLAB IDE, but can
be used stand-alone.