Working version 1.0 directly replaces the stock Vent-Axia firmware.
If SET button is held then "ENTER COMMISSION" is displayed. Whether the
mode is actually entered is still rather hit-and-miss but it does work
sometimes. More experimentation with the COMMISSION_BUTTON_DELAY value
is probably required.

Otherwise all buttons repeat when long-pressed to allow special functions
as in the stock firmware.

"Tool" scripts to allow the use of a PICKit3 programmer in MikroC PRO
are included (no need for MPLAB if you don't want it or don't have the
3GB that it takes up).